refactor: Move types to separate file and clean up imports
