+++ src/app/nonce/components/sidebar.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { NonceRecord } from "../mock-data";
+import { NonceRecord } from "../types";
 import { Filter, BarChart2, ArrowLeftRight, Calendar as CalendarIcon } from "lucide-react";
 import { cn } from "@/lib/utils";
 import { useEffect, useState } from "react";
@@ -8,14 +8,14 @@ import { DateRange } from "react-day-picker";
 import { format, subDays, addDays, subYears } from "date-fns";
 import { DateRangePicker } from "./date-range-picker";
 import { TimeComparisonSelector, ComparisonOption } from "./time-comparison-selector";
-import { 
-  createDefaultQuery, 
-  getAvailableMeasures, 
-  getAvailableDimensions, 
-  getAvailableFarmNames, 
-  updateMeasures, 
-  updateDimensions, 
-  updateFilters, 
+import {
+  createDefaultQuery,
+  getAvailableMeasures,
+  getAvailableDimensions,
+  getAvailableFarmNames,
+  updateMeasures,
+  updateDimensions,
+  updateFilters,
   updateDateRange,
   ExtendedQuery
 } from "../utils/cube-query-builder";
@@ -35,7 +35,7 @@ export function Sidebar({ nonceData, onQueryStateChange, onComparisonChange }: S
   const measures = getAvailableMeasures();
   const dimensions = getAvailableDimensions();
   const farmNames = getAvailableFarmNames();
-  
+
 
   // Group measures by folder
   const measuresByFolder = measures.reduce((acc, measure) => {
@@ -58,7 +58,7 @@ export function Sidebar({ nonceData, onQueryStateChange, onComparisonChange }: S
       onQueryStateChange(queryState);
     }
   }, [queryState, onQueryStateChange]);
-  
+
   // Pass comparison selection to parent when it changes
   useEffect(() => {
     if (onComparisonChange) {
@@ -72,10 +72,10 @@ export function Sidebar({ nonceData, onQueryStateChange, onComparisonChange }: S
       .filter(([folderName]) => folderName !== folder)
       .flatMap(([_, items]) => items.map(item => item.id))
       .filter(id => currentMeasures.includes(id));
-    
+
     // Combine with newly selected measures from this folder
     const newMeasures = [...otherFolderMeasures, ...selectedIds];
-    
+
     setQueryState(updateMeasures(queryState, newMeasures));
   };
 
@@ -87,9 +87,9 @@ export function Sidebar({ nonceData, onQueryStateChange, onComparisonChange }: S
   // Handle farm filter selection
   const handleFarmChange = (farms: string[]) => {
     setQueryState(updateFilters(
-      queryState, 
-      'metrics.workspace_name', 
-      'equals', 
+      queryState,
+      'metrics.workspace_name',
+      'equals',
       farms
     ));
   };
@@ -99,7 +99,7 @@ export function Sidebar({ nonceData, onQueryStateChange, onComparisonChange }: S
     if (range?.from && range?.to) {
       const formattedFrom = format(range.from, 'yyyy-MM-dd');
       const formattedTo = format(range.to, 'yyyy-MM-dd');
-      
+
       setQueryState(updateDateRange(queryState, [formattedFrom, formattedTo]));
     }
   };
@@ -110,9 +110,9 @@ export function Sidebar({ nonceData, onQueryStateChange, onComparisonChange }: S
         <div className="pb-2 mb-2">
           <h3 className="text-base font-medium">Hashing</h3>
         </div>
-        
-        <Category 
-          title="Measure" 
+
+        <Category
+          title="Measure"
           icon={<BarChart2 className="h-4 w-4" />}
           defaultOpen={true}
         >
@@ -122,7 +122,7 @@ export function Sidebar({ nonceData, onQueryStateChange, onComparisonChange }: S
               const selectedIds = items
                 .filter(item => currentMeasures.includes(item.id))
                 .map(item => item.id);
-              
+
               return (
                 <div key={folder} className="space-y-1">
                   <div className="font-medium">{folder}</div>
@@ -139,9 +139,9 @@ export function Sidebar({ nonceData, onQueryStateChange, onComparisonChange }: S
             })}
           </div>
         </Category>
-        
-        <Category 
-          title="Breakdown" 
+
+        <Category
+          title="Breakdown"
           icon={<ArrowLeftRight className="h-4 w-4" />}
         >
           <div className="space-y-2">
@@ -153,31 +153,35 @@ export function Sidebar({ nonceData, onQueryStateChange, onComparisonChange }: S
             />
           </div>
         </Category>
-        
-        <Category 
-          title="Filter" 
-          icon={<Filter className="h-4 w-4" />}
-        >
-          <div className="space-y-2">
-            <CheckboxGroup
-              title="Farm Name"
-              items={farmNames.map(farm => ({ 
-                id: farm, 
-                name: farm 
-              }))}
-              selectedItems={queryState.filters
-                ? (queryState.filters.find((filter: any) => 
-                    (filter as BinaryFilter).member === 'metrics.workspace_name'
-                  ) as BinaryFilter | undefined)?.values || []
-                : []}
-              
-              onChange={handleFarmChange}
-            />
-          </div>
-        </Category>
-        
-        <Category 
-          title="Compare" 
+
+        {
+          (queryState?.dimensions || [])?.length > 0 && (
+            <Category
+              title="Filter"
+              icon={<Filter className="h-4 w-4" />}
+            >
+              <div className="space-y-2">
+                <CheckboxGroup
+                  title="Farm Name"
+                  items={farmNames.map(farm => ({
+                    id: farm,
+                    name: farm
+                  }))}
+                  selectedItems={queryState.filters
+                    ? (queryState.filters.find((filter: any) =>
+                      (filter as BinaryFilter).member === 'metrics.workspace_name'
+                    ) as BinaryFilter | undefined)?.values || []
+                    : []}
+
+                  onChange={handleFarmChange}
+                />
+              </div>
+            </Category>
+          )
+        }
+
+        <Category
+          title="Compare"
           icon={<ArrowLeftRight className="h-4 w-4" />}
         >
           <div className="space-y-4">
@@ -188,8 +192,8 @@ export function Sidebar({ nonceData, onQueryStateChange, onComparisonChange }: S
           </div>
         </Category>
 
-        <Category 
-          title="Time Range" 
+        <Category
+          title="Time Range"
           icon={<CalendarIcon className="h-4 w-4" />}
         >
           <div className="space-y-2">
